server/requirements.txt
cairosvg
python-chess
pandas
numpy
tensorflow
Flask
Flask-RESTful
python-dotenv